+++ loopback-dropwizard/src/main/java/com/flipkart/loopback/resources/DWResource.java
@@ -115,7 +115,7 @@ public interface DWResource<T extends PersistedModel> {
    */
   @GET
   @Path("/count")
-  public int count(@QueryParam("filter") Filter filter, @Context ContainerRequestContext
+  public long count(@QueryParam("filter") Filter filter, @Context ContainerRequestContext
       requestContext);
 
   /*
